git commit "send exercises/touch server"

+++ app/src/main/java/com/movetoplay/presentation/ui/child_home/ExercisesPerformedOnDay.kt
@@ -69,7 +69,7 @@ fun ExercisesPerformedOnDay(
                 .background(MaterialTheme.colorScheme.primary, RoundedCornerShape(10.dp))
         ) {
             val jumps = Pref.jumps
-            val progressJumping = jumps.toInt()
+            val progressJumping = jumps
             Box(
                 modifier = Modifier
                     .width(300.dp)
@@ -118,7 +118,7 @@ fun ExercisesPerformedOnDay(
                 .background(MaterialTheme.colorScheme.primary, RoundedCornerShape(10.dp))
         ) {
             val pushups = Pref.push_ups
-            val progressPushUps = pushups.toInt()
+            val progressPushUps = pushups
             Box(
                 modifier = Modifier
                     .width(300.dp)
@@ -168,7 +168,7 @@ fun ExercisesPerformedOnDay(
                 .background(MaterialTheme.colorScheme.primary, RoundedCornerShape(10.dp))
         ) {
             val squats = Pref.sits
-            val progressSquats = squats.toInt()
+            val progressSquats = squats
             Box(
                 modifier = Modifier
                     .width(300.dp)
